# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Bairros Fortaleza",
    "description": "Some description..."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
    "/bairro/{id}": {
      "get": {
        "description": "Recebe o id de um bairro e retorna dados básicos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/localizar/numero/{numero}/cep/{cep}": {
      "get": {
        "description": "Recebe um endereço (número e cep) e retorna coordenadas.",
        "parameters": [
          {
            "name": "numero",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cep",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/pesquisar": {
      "get": {
        "description": "Recebe um CEP e retorna dados básicos sobre o bairro em que está localizado.",
        "parameters": [
          {
            "name": "cep",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/comparar": {
      "post": {
        "description": "Recebe três endereços (número e CEP) e retorna dados básicos, \r \t\tresultados das comparações entre os indicadores e as distâncias entre as localidades.",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/comparadorSchema2"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/comparadorSchema2"
              }
            }
          }
        }
      }
    },
    "/calcular-distancia": {
      "post": {
        "description": "Recebe dois endereços (número e CEP) e retorna as distâncias entre as localidades.",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/comparadorSchema"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/comparadorSchema"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "comparadorSchema": {
        "type": "object",
        "properties": {
          "origem": {
            "type": "object",
            "properties": {
              "numero": {
                "type": "string",
                "example": "1321"
              },
              "cep": {
                "type": "string",
                "example": "60811-905"
              }
            },
            "required": [
              "numero",
              "cep"
            ]
          },
          "destino": {
            "type": "object",
            "properties": {
              "numero": {
                "type": "string",
                "example": "71"
              },
              "cep": {
                "type": "string",
                "example": "60421440"
              }
            },
            "required": [
              "numero",
              "cep"
            ]
          }
        },
        "required": [
          "origem",
          "destino"
        ],
        "xml": {
          "name": "comparadorSchema"
        }
      },
      "comparadorSchema2": {
        "type": "object",
        "properties": {
          "origem": {
            "type": "object",
            "properties": {
              "numero": {
                "type": "string",
                "example": "1321"
              },
              "cep": {
                "type": "string",
                "example": "60811-905"
              }
            },
            "required": [
              "numero",
              "cep"
            ]
          },
          "destino": {
            "type": "object",
            "properties": {
              "numero": {
                "type": "string",
                "example": "71"
              },
              "cep": {
                "type": "string",
                "example": "60421440"
              }
            },
            "required": [
              "numero",
              "cep"
            ]
          },
          "localDeInteresse": {
            "type": "object",
            "properties": {
              "numero": {
                "type": "string",
                "example": "282"
              },
              "cep": {
                "type": "string",
                "example": "60055402"
              }
            },
            "required": [
              "numero",
              "cep"
            ]
          }
        },
        "required": [
          "origem",
          "destino",
          "localDeInteresse"
        ],
        "xml": {
          "name": "comparadorSchema2"
        }
      },
      "bairroId": {
        "type": "object",
        "properties": {
          "id_pmf": {
            "type": "string",
            "example": "77"
          },
          "bairro_nome": {
            "type": "string",
            "example": "Rachel de Queiroz"
          },
          "indicador": {
            "type": "object",
            "properties": {
              "Socioeconômico": {
                "type": "object",
                "properties": {
                  "IDH Renda": {
                    "type": "number",
                    "example": 0.115
                  },
                  "IDH": {
                    "type": "number",
                    "example": 0.181
                  },
                  "IDH Educação": {
                    "type": "number",
                    "example": 0.948
                  }
                }
              }
            }
          }
        },
        "xml": {
          "name": "bairroId"
        }
      }
    }
  }
}